ArtPark_CodeForge/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.28.1
networkx==3.4.2
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
